requirements.txt
Flask==2.2.2
requests==2.28.1
Bootstrap-Flask==2.0.0
alpaca-trade-api==2.3.0
